Changes to be committed:     modified:   software/o_c_REV/HEM_RunglBook.ino 	modified:   software/o_c_REV/platformio.ini New app: 	software/o_c_REV/HEM_Relabi.ino This branch of Benispheres contains a redesigned Runglbook app with a second CV input for changing the bit depth of the output from 2 to 8 bits. It's second output provides new trigger every time a new CV value is generated at the first output. The encoder can be used to select a new bitdepth.
A new app has been added for generating Relabi waves. A relabi wave
is a deterministic method of creating CV that feels random but also
feels related to the tempo of its constituent LFOs. The app currently
doesn't use the inputs. It does output the relabi wave and a gate based
on a threshold set by the encoder. The gate provides pulses that feel
like a rhythm that is always slipping the pulse, as you can read about
on John Berndt's website. The encoder is also used to select the
controls for four different LFOs and used to set their frequencies and
cross modulation.

Future plans include setting the phase of each LFO and the ability
to send a trigger into the O_c to reset the LFOs to their assigned
phases.